made some cleanup in code for semver
