+++ splitio/sync/segment.py
@@ -144,13 +144,13 @@ class SegmentSynchronizer(object):
         :return: True if no error occurs. False otherwise.
         :rtype: bool
         """
-        fetch_options = FetchOptions(True, None, None, None)  # Set Cache-Control to no-cache
+        fetch_options = FetchOptions(True, spec=None)  # Set Cache-Control to no-cache
         successful_sync, remaining_attempts, change_number = self._attempt_segment_sync(segment_name, fetch_options, till)
         attempts = _ON_DEMAND_FETCH_BACKOFF_MAX_RETRIES - remaining_attempts
         if successful_sync:  # succedeed sync
             _LOGGER.debug('Refresh completed in %d attempts.', attempts)
             return True
-        with_cdn_bypass = FetchOptions(True, change_number, None, None)  # Set flag for bypassing CDN
+        with_cdn_bypass = FetchOptions(True, change_number, spec=None)  # Set flag for bypassing CDN
         without_cdn_successful_sync, remaining_attempts, change_number = self._attempt_segment_sync(segment_name, with_cdn_bypass, till)
         without_cdn_attempts = _ON_DEMAND_FETCH_BACKOFF_MAX_RETRIES - remaining_attempts
         if without_cdn_successful_sync:
